requirements.txt
# ✅ requirements.txt

# Core LangChain agent stack
langchain==0.1.20
langchain-community==0.0.35
langchain-google-genai==0.0.10

# LLMs + Google Gemini support
google-generativeai==0.5.4

# FAISS & Embeddings
faiss-cpu==1.7.4
sentence-transformers==2.7.0

# Neo4j for knowledge graph
neo4j==5.19.0

# Chat UI
gradio==4.21.0

# Utilities
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4

# Optional: for reranking
transformers==4.41.1
scikit-learn==1.5.0